backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-environ>=0.11
psycopg2-binary>=2.9
django-cors-headers>=4.0
gunicorn>=21.2